cleaned up PR
