+++ backend/routes/comments.routes.js
@@ -9,7 +9,6 @@ router.route('/:id').get(authenticateUser, getAllCommentsController)
 
 router.route('/').post(authenticateUser, createCommentController)
 
-//delete, edit and others can come later
-//shouldn't be too hard can just fully replace the comment part
+//TODO: delete & edit
 
 module.exports = router
